Rename gpt_index/readers/google to fix import path mangling of google.protobuf
